pyproject.toml
[project]
name = "kaminai"
version = "0.1.0"
description = "KaminAI - Personal blog and portfolio website"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # FastAPI
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sse-starlette>=2.0.0",

    # MongoDB
    "motor>=3.3.0",
    "pymongo>=4.6.0",

    # HTTP Client
    "httpx>=0.26.0",

    # Utilities
    "python-dotenv>=1.0.0",

    # Authentication
    "PyJWT>=2.8.0",

    # File uploads
    "python-multipart>=0.0.6",

    # AI/LangGraph
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "faiss-cpu>=1.8.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
